feat(Seneca): implement all actions as seneca services
Each task file has become a seneca microservice file that adds the
appropriate action using the pattern
{role:Tasks,cmd:update/create/remove}

The dispatcher now calls seneca.act to route to the correct action. Some
changes to the layout of action calls needed to be made, mainly around
create and update actions that have a "values" object. This change is
also reflected in the frontend app.

Authorization is implemented using .wrap around some of the actions.